Factor does not need to be an integer
git-svn-id: file:///home/projects/mule/scm@11726 bf997673-6b11-0410-b953-e057580c5b09

+++ branches/mule-2.0.x/core/src/main/java/org/mule/transformer/simple/SimpleMathTransformer.java
@@ -16,7 +16,7 @@ import org.mule.transformer.AbstractTransformer;
 import org.mule.util.NumberUtils;
 
 /** 
- * A simple transformer which adds/subtracts/multiplies/divides a constant factor to integer messages. 
+ * A simple transformer which adds/subtracts/multiplies/divides a constant factor to numeric messages. 
  */
 public class SimpleMathTransformer extends AbstractTransformer 
 {
@@ -24,7 +24,7 @@ public class SimpleMathTransformer extends AbstractTransformer
     private String operation = "add";
     
     /** Factor to be applied */
-    private int factor;
+    private double factor;
     
     public SimpleMathTransformer()
     {
@@ -72,12 +72,12 @@ public class SimpleMathTransformer extends AbstractTransformer
         this.operation = operation;
     }
 
-    public int getFactor()
+    public double getFactor()
     {
         return factor;
     }
 
-    public void setFactor(int factor)
+    public void setFactor(double factor)
     {
         this.factor = factor;
     }
